BackendWithFlask/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Flask-Mail==0.10.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
PyJWT==2.8.0
kafka-python==2.2.15
MarkupSafe==3.0.2
mode==4.3.2
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
opentracing==1.3.0
pandas==2.3.3
propcache==0.3.2
psycopg2-binary==2.9.11
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.44
terminaltables==3.1.10
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
venusian==1.2.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
Authlib~=1.6.5
elasticsearch~=9.1.1
dotenv~=0.9.9
gunicorn>=22.0.0
gevent>=24.2.1